bit of refactoring in the lib directory and some more featuremap annotations
